Insane for loop
Between lines 81-155; there are two functions that control user input and assigning outputs on display-board.
The craziness happens when a user chooses an already used tile and when you get to pick a new unused tile, the program doesn't accept the input and you can't proceed with the program. 
I looked at through the debug tool and i witnessed the strangest act of all acts.. 
Long story short: The program goes through the double for loop like it should but when it is about to end the loop, it some how magically ends up in a else statement in the middle, with a variable which was incrementing by one through out the loop is now something entirely different.. It's fun, weird but fun.